+++ lib/core/api/services/upload_service/upload_service.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/foundation.dart';
+import 'package:http/http.dart';
+
+import 'upload_service_impl.dart';
+
+abstract class UploadService {
+  upload(MultipartFile multipart,  {
+    @required OnUploadProgressCallback onProgressCallback
+  });
+}